requirements.txt
Flask
SpeechRecognition
pyttsx3
wikipedia-api
pyaudio
gunicorn